+++ helpers/user_helper.py
@@ -74,72 +74,3 @@ def set_country(user_id: int, country_code: str) -> None:
         "UPDATE users SET country = %s WHERE id = %s LIMIT 1",
         (country_code, user_id),
     )
-
-
-def insert_ban_log(
-    user_id: int,
-    summary: str,
-    detail: str,
-    prefix: bool = True,
-    from_id: Optional[int] = None,
-) -> None:
-    """Inserts a ban log for a user into the database.
-
-    Args:
-        user_id (int): The ID of the user to assign the log to.
-        summary (str): A short description of the reason.
-        detail (str): A more detailed, in-depth description of the reason.
-        prefix (bool, optional): Whether the detail should be prefixed by
-            the peppy signature. Defaults to True.
-        from_id (int, optional): The ID of the user who banned the user.
-            Defaults to the configured bot.
-    """
-
-    if from_id is None:
-        from_id = config.SRV_BOT_ID
-
-    if prefix:
-        detail = "pep.py Autoban: " + detail
-
-    glob.db.execute(
-        "INSERT INTO ban_logs (from_id, to_id, summary, detail) VALUES (%s, %s, %s, %s)",
-        (
-            from_id,
-            user_id,
-            summary,
-            detail,
-        ),
-    )
-
-
-def restrict_with_log(
-    user_id: int,
-    summary: str,
-    detail: str,
-    prefix: bool = True,
-    from_id: Optional[int] = None,
-) -> None:
-    """Restricts the user alongside inserting a log into the database.
-
-    Args:
-        user_id (int): The ID of the user to assign the log to.
-        summary (str): A short description of the reason.
-        detail (str): A more detailed, in-depth description of the reason.
-        prefix (bool, optional): Whether the detail should be prefixed by
-            the peppy signature. Defaults to True.
-        from_id (int, optional): The ID of the user who banned the user.
-            Defaults to the configured bot.
-    """
-
-    if from_id is None:
-        from_id = config.SRV_BOT_ID
-
-    glob.db.execute(
-        f"UPDATE users SET privileges = privileges & ~{privileges.USER_PUBLIC}, "
-        "ban_datetime = UNIX_TIMESTAMP() WHERE id = %s LIMIT 1",
-        (user_id,),
-    )
-    glob.redis.publish("peppy:ban", user_id)
-    removeFromLeaderboard(user_id)
-
-    insert_ban_log(user_id, summary, detail, prefix, from_id)
